+++ src/main/java/ru/l3r8y/complaint/WrongMethodSignature.java
@@ -21,37 +21,37 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-
 package ru.l3r8y.complaint;
 
-import lombok.RequiredArgsConstructor;
-import ru.l3r8y.ClassName;
+import lombok.AllArgsConstructor;
 import ru.l3r8y.Complaint;
+import ru.l3r8y.Method;
 
 /**
- * Complaint for naming issues.
+ * Complaint about wrong method signature.
  *
- * @since 0.1.6
+ * @since 0.1.0
  */
-@RequiredArgsConstructor
-public final class WrongClassNamingComplaint implements Complaint {
+@AllArgsConstructor
+public class WrongMethodSignature implements Complaint {
 
     /**
-     * Class with bad naming.
+     * The method.
      */
-    private final ClassName clazz;
+    private final Method method;
 
     /**
-     * The explanation.
+     * The explanation of what was wrong.
      */
     private final String explanation;
 
     @Override
-    public String message() {
+    public final String message() {
         return String.format(
-            "'%s': Class '%s' has bad naming, %s\n",
-            this.clazz.path().toString(),
-            this.clazz.value(),
+            "'%s': Method '%s#%s' has wrong method signature, because %s",
+            this.method.path(),
+            this.method.className(),
+            this.method.name(),
             this.explanation
         );
     }
