+++ src/main/java/ru/l3r8y/rule/MethodChangesState.java
@@ -33,7 +33,7 @@ import lombok.AllArgsConstructor;
 import ru.l3r8y.Complaint;
 import ru.l3r8y.Method;
 import ru.l3r8y.Rule;
-import ru.l3r8y.complaint.WrongMethodSignatureComplaint;
+import ru.l3r8y.complaint.WrongMethodSignature;
 
 /**
  * It checks if a method contains an assignment.
@@ -41,7 +41,7 @@ import ru.l3r8y.complaint.WrongMethodSignatureComplaint;
  * @since 0.1.0
  */
 @AllArgsConstructor
-public final class MethodContainsAssigment implements Rule {
+public final class MethodChangesState implements Rule {
 
     /**
      * Pattern which matches this assignments.
@@ -60,7 +60,7 @@ public final class MethodContainsAssigment implements Rule {
     public Collection<Complaint> complaints() {
         return new ConditionRule(
             this::containsAssigment,
-            new WrongMethodSignatureComplaint(
+            new WrongMethodSignature(
                 this.method,
                 "method body contains an assignment, setters violates OOP principles"
             )
@@ -73,6 +73,6 @@ public final class MethodContainsAssigment implements Rule {
      * @return A boolean value.
      */
     private boolean containsAssigment() {
-        return MethodContainsAssigment.PATTERN.matcher(this.method.body()).find();
+        return MethodChangesState.PATTERN.matcher(this.method.body()).find();
     }
 }
